+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 ## ChangeLog
 
+* Added support for [New Relic Browser](https://docs.newrelic.com/docs/browser)
+  using the new `BrowserTimingHeader` method on the
+  [`Transaction`](https://godoc.org/github.com/newrelic/go-agent#Transaction)
+  which returns a
+  [BrowserTimingHeader](https://godoc.org/github.com/newrelic/go-agent#BrowserTimingHeader).
+  The New Relic Browser JavaScript code measures page load timing, also known as
+  real user monitoring.  The Pro version of this feature measures AJAX requests,
+  single-page applications, JavaScript errors, and much more!  Example use:
+
+```go
+func browser(w http.ResponseWriter, r *http.Request) {
+	hdr, err := w.(newrelic.Transaction).BrowserTimingHeader()
+	if nil != err {
+		log.Printf("unable to create browser timing header: %v", err)
+	}
+	// BrowserTimingHeader() will always return a header whose methods can
+	// be safely called.
+	if js := hdr.WithTags(); js != nil {
+		w.Write(js)
+	}
+	io.WriteString(w, "browser header page")
+}
+```
+
 ## 2.4.0
 
 * Introduced `Transaction.Application` method which returns the `Application`